refactor: CarHistory를 도메인 영역으로 옮김
